+++ contracts/interfaces/IDrawCalculator.sol
@@ -3,6 +3,6 @@ pragma solidity 0.8.6;
 import "hardhat/console.sol";
 
 interface IDrawCalculator {
-  function calculate(address user, uint256[] calldata randomNumbers, uint256[] calldata timestamps, uint256[] calldata prizes, bytes calldata data) external returns (uint256);
+  function calculate(address user, uint256[] calldata randomNumbers, uint32[] calldata timestamps, uint256[] calldata prizes, bytes calldata data) external returns (uint256);
   
 }
